bugfixes, thanks theo77186 for pointing them

+++ src/search.cc
@@ -27,15 +27,15 @@ void Search::init_LMR_array(){
 
   for (int i = 0; i < 34; i++){
     for (int j = 0; j < 34; j++){
-      _lmr_R_array[i][j] = 0.1 + (pow(i, 0.15) * pow(j, 0.15))/1.75;
+      _lmr_R_array[i][j] = (int) (0.1 + (pow(i, 0.15) * pow(j, 0.15))/1.75);
     }
   }
   // 2. Initialization of the LMP array.
   // Current formula is completely based on the 
   // Weiss chess engine.
   for (int i = 0; i < 99; i++){
-    _lmp_Array[i][0] = (3 + pow( i, 2) * 2) / 2;
-    _lmp_Array[i][1] = (3 + pow( i, 2) * 2);
+    _lmp_Array[i][0] = (int) ((3 + pow( i, 2) * 2) / 2);
+    _lmp_Array[i][1] = (int) (3 + pow( i, 2) * 2);
   }
 
 }
